chore(redis): rename module.ts -> plugin.ts
