+++ plugins/app-redis/src/redis-cli-panel/plugin.test.ts
@@ -1,5 +1,5 @@
 import { PanelPlugin } from '@grafana/data';
-import { plugin } from './module';
+import { plugin } from './plugin';
 
 /*
  Plugin
