+++ plugins/app-redis/src/redis-cli-panel/plugin.ts
