+++ plugins/app-redis/src/redis-cpu-panel/plugin.ts
