+++ plugins/app-redis/src/redis-gears-panel/plugin.ts
