+++ plugins/app-redis/src/redis-keys-panel/plugin.test.ts
@@ -1,21 +1,17 @@
 import { PanelPlugin } from '@grafana/data';
-import { plugin } from './module';
+import { plugin } from './plugin';
 
 /**
- * CPU Panel
+ * Keys Panel
  */
-describe('RedisCPUPanel', () => {
+describe('RedisKeysPanel', () => {
   it('Should be instance of PanelPlugin', () => {
     expect(plugin).toBeInstanceOf(PanelPlugin);
   });
 
   it('Should add interval input', () => {
-    /**
-     * Builder
-     */
     const builder: any = {
       addSliderInput: jest.fn().mockImplementation(() => builder),
-      addRadio: jest.fn().mockImplementation(() => builder),
     };
 
     plugin['optionsSupplier'](builder);
