+++ plugins/app-redis/src/redis-keys-panel/plugin.ts
