+++ plugins/app-redis/src/redis-latency-panel/plugin.ts
